refactor: restructed file-path of Search-components
Moved Components from /Shared/Filtering to /Shared/Search
